test: 기능 구현 코드 리팩터링으로 인한 테스트 코드 수정

+++ src/main/java/roomescape/repository/JdbcMemberDao.java
@@ -42,7 +42,7 @@ public class JdbcMemberDao implements MemberRepository {
                 .addValue("email", registrationDetails.email())
                 .addValue("password", registrationDetails.password());
 
-        jdbcInsert.executeAndReturnKey(params).longValue();
+        jdbcInsert.executeAndReturnKey(params);
     }
 
     @Override
